server/requirement.txt
# Core framework
Django>=5.0,<6.0
djangorestframework>=3.15.0

# Authentication & JWT
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.1

# Database
psycopg2-binary>=2.9.9     # if using PostgreSQL (Supabase)
django-environ>=0.11.2

# CORS for frontend communication
django-cors-headers>=4.4.0

# AI / LLM integrations (optional)
openai>=1.40.0              # or google-generativeai / litellm etc.

# Utilities
requests>=2.32.3

# For async tasks / websocket (future optional)
channels>=4.1.0
